tutorials/delivering-cloud-monitoring-notifications-to-third-party-services/scripts/requirements.txt
google-cloud-monitoring==1.0.0
google-cloud-secret-manager==1.0.0
pycryptodome==3.6.6
requests-oauthlib==1.3.0
oauthlib==3.1.0
